requirements.txt
Flask==3.0.0
pdfplumber==0.11.0
PyMuPDF==1.23.8
reportlab==4.0.7
python-dotenv==1.0.0
openai==1.54.0
Pillow==10.1.0
pytesseract==0.3.10
httpx==0.27.0
Werkzeug==3.0.0